Rewrite of windowing system
This is a rewrite of the way glk windows are handled internally. The Glk::*Widget classes handle the Qt/GUI aspects and the Glk::*Window classes handle the glk logic. These two talk between each other via the Glk::*WindowController classes.

This results in overall performance improvements, as well as less flickering.

Hyperlinks and some input events are not yet reimplemented.